Bugfix: missing last cycle
We check if the last cycle has complete, if not it is removed from
analysed results.
This doesn't work with Neware or EC-lab harvested data, because we
don't know if the job is finished or not.
Now this is only applied when the job status is known.

Remove voltage limit TODO - this won't work because the voltage can
change very rapidly especially at low capacities.